+++ release_notes.md
@@ -22,6 +22,7 @@ the training graph now.
 * Fix #275 : better error support when reconnecting hierarchical graph with wrong ports
 * Fix #19 Fix #39 : new menu popup action to check liveness of PiSDF graphs.
 * Fix #322 : as C/C++ parsing is now done by Eclipse CDT, we accept headers with .h .H .hxx .hpp .hh .h++ file extensions, but they are not yet supported by regular codegen.
+* Fix #165 : UI warns if fifo types do not correspond to actor refinement using them.
 
 
 ## Release version 3.21.0